+++ .dev-doc/main/CHANGELOG.md
@@ -1,1 +1,4 @@
 # Changelog
+
+## 2026-05-31
+- 14:45 fix: ISSUE-I001：iterate token 未绑定命令参数